feat: 태그 API는 훅만 담당하도록 수정, 어댑터 등에서 mock 요소 제거(#163)
